fix: clay laser interface doesn't require synchro parts
